// apps/candyland/src/app/models/controlsName.ts
export enum ControlsName {
  name = 'name',
  headlineMessage = 'headlineMessage',
  subHeadlineMessage = 'subHeadlineMessage',
  buttonText = 'buttonText',
  background = 'background',

  pinata = 'pinata',
  cardBackground = 'cardBackground',

  treeType = 'treeType',
  giftBox = 'giftBox',
  gameGift = 'gameGift',


}